refactor: move reflect feature from Shaderity.ts to Reflection.ts
